README.md
My capstone project is an interactive, narrative-driven experience that immerses users in a haunting narrative. Choose your archetype—Detective or Entity—and uncover secrets hidden within a dark, surreal world. There will be hidden triggers and elements within the story and it's your job to uncover them as the story is told. Headphones recommended.